pyproject.toml
[tool.poetry]
name = "obyn"
version = "0.1.0"
description = ""
authors = ["terabyte3 <terry.bazemoreiii@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.10"
discord-py-interactions = "^4.2.1"
aiohttp = "^3.8.1"
cchardet = "^2.1.7"
aiodns = "^3.0.0"
pydantic = "^1.9.1"
orjson = "^3.7.2"
python-dotenv = "^0.20.0"
disnake = "^2.5.1"
jishaku = "^2.5.0"
asteval = "^0.9.27"
humanize = "^4.1.0"
"betterstatcord.py" = "^3.5.0"
cachetools = "^5.2.0"

[tool.poetry.dev-dependencies]
types-cachetools = "^5.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
